Fix a few issues in the Monitoring module

+++ src/monitoring/CPUMonitor.hpp
@@ -52,7 +52,7 @@ public:
 	//! \param[in] virtualCPUId The identifier of the CPU
 	inline void cpuBecomesActive(int virtualCPUId)
 	{
-		assert(virtualCPUId <= _numCPUs);
+		assert((size_t) virtualCPUId <= _numCPUs);
 
 		_cpuStatistics[virtualCPUId].cpuBecomesActive();
 	}
@@ -62,7 +62,7 @@ public:
 	//! \param[in] virtualCPUId The identifier of the CPU
 	inline void cpuBecomesIdle(int virtualCPUId)
 	{
-		assert(virtualCPUId <= _numCPUs);
+		assert((size_t) virtualCPUId <= _numCPUs);
 
 		_cpuStatistics[virtualCPUId].cpuBecomesIdle();
 	}
@@ -72,7 +72,7 @@ public:
 	//! \param[in] virtualCPUId The identifier of the CPU
 	inline float getActiveness(int virtualCPUId)
 	{
-		assert(virtualCPUId < _numCPUs);
+		assert((size_t) virtualCPUId < _numCPUs);
 
 		return _cpuStatistics[virtualCPUId].getActiveness();
 	}
